requirements.txt
fastapi
uvicorn[standard]
python-multipart
pyjwt
bcrypt